# processor/internal/gateway/docs/swagger.yaml
basePath: /
definitions:
  handler.APIError:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  pb.Account:
    properties:
      account_id:
        description: Unique identifier for the trading account.
        type: string
      account_type:
        description: Type of account (e.g., Stock, Future).
        type: string
      broker_id:
        description: Identifier for the broker.
        type: string
      person_id:
        description: Unique identifier for the account holder.
        type: string
      signed:
        description: Whether the account has signed the trading agreement.
        type: boolean
      username:
        description: User's name associated with the account.
        type: string
    type: object
  pb.AccountBalance:
    properties:
      acc_balance:
        description: Current account balance.
        type: number
      date:
        description: Date for which the balance is reported.
        type: string
      errmsg:
        description: Error message if the fetch failed.
        type: string
      status:
        allOf:
        - $ref: '#/definitions/pb.FetchStatus'
        description: Success/Failure status of the fetch operation.
    type: object
...
  pb.DailyQuotes:
    properties:
      amount:
        description: Daily amount.
        items:
          type: integer
        type: array
      close:
        description: Daily close.
        items:
          type: number
        type: array
      code:
        description: Security codes.
        items:
          type: string
        type: array
      date:
        description: Report dates.
        items:
          type: string
        type: array
      high:
        description: Daily high.
        items:
          type: number
        type: array
      low:
        description: Daily low.
        items:
          type: number
        type: array
      open:
        description: Daily open.
        items:
          type: number
        type: array
      transaction:
        description: Transaction count.
        items:
          type: integer
        type: array
      volume:
        description: Daily volume.
        items:
          type: integer
        type: array
    type: object
...
  pb.Empty:
    type: object
...
  pb.FetchContractsRequest:
    properties:
      contract_download:
        description: Whether to perform a full download.
        type: boolean
    type: object
  pb.FetchStatus:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-comments:
      FetchStatus_FETCH_STATUS_FAIL: Fail. 失敗
      FetchStatus_FETCH_STATUS_SUCCESS: Success. 成功
    x-enum-descriptions:
    - ""
    - Success. 成功
    - Fail. 失敗
    x-enum-varnames:
    - FetchStatus_FETCH_STATUS_UNSPECIFIED
    - FetchStatus_FETCH_STATUS_SUCCESS
    - FetchStatus_FETCH_STATUS_FAIL
...
  pb.GetDailyQuotesRequest:
    properties:
      date:
        description: YYYY-MM-DD.
        type: string
    type: object
...
  pb.GetKbarsRequest:
    properties:
      contract_code:
        type: string
      end_date:
        description: YYYY-MM-DD.
        type: string
      start_date:
        description: YYYY-MM-DD.
        type: string
    type: object
  pb.GetMarginRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
    type: object
...
  pb.GetScannersRequest:
    properties:
      ascending:
        description: Sort direction.
        type: boolean
      count:
        description: Max number of results.
        type: integer
      date:
        description: Date for scanning.
        type: string
      scanner_type:
        allOf:
        - $ref: '#/definitions/pb.ScannerType'
        description: Ranking criteria (e.g., volume, change).
    type: object
  pb.GetScannersResponse:
    properties:
      scanners:
        items:
          $ref: '#/definitions/pb.ScannerItem'
        type: array
    type: object
  pb.GetSettlementsRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
    type: object
  pb.GetSettlementsResponse:
    properties:
      settlements:
        items:
          $ref: '#/definitions/pb.Settlement'
        type: array
    type: object
...
  pb.GetSnapshotsRequest:
    properties:
      contract_codes:
        description: List of security codes.
        items:
          type: string
        type: array
    type: object
  pb.GetSnapshotsResponse:
    properties:
      snapshots:
        items:
          $ref: '#/definitions/pb.Snapshot'
        type: array
    type: object
...
  pb.GetTicksRequest:
    properties:
      contract_code:
        type: string
      date:
        description: Date YYYY-MM-DD.
        type: string
    type: object
  pb.GetTradingLimitsRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
    type: object
  pb.Kbars:
    properties:
      amount:
        description: Total period amounts.
        items:
          type: number
        type: array
      close:
        description: Close prices.
        items:
          type: number
        type: array
      high:
        description: High prices.
        items:
          type: number
        type: array
      low:
        description: Low prices.
        items:
          type: number
        type: array
      open:
        description: Open prices.
        items:
          type: number
        type: array
      ts:
        description: Period timestamps.
        items:
          type: integer
        type: array
      volume:
        description: Total period volumes.
        items:
          type: integer
        type: array
    type: object
  pb.ListAccountsResponse:
    properties:
      accounts:
        items:
          $ref: '#/definitions/pb.Account'
        type: array
    type: object
...
  pb.Margin:
    properties:
      available_margin:
        description: Margin available for new trades.
        type: number
      collateral_amount:
        description: Cash collateral amount.
        type: number
      deposit_withdrawal:
        description: Net deposits and withdrawals.
        type: number
      equity:
        description: Total equity.
        type: number
      equity_amount:
        description: Liquid equity amount.
        type: number
      fee:
        description: Total fees paid.
        type: number
      future_open_position:
        description: Number of open futures contracts.
        type: number
      future_settle_profitloss:
        description: Settlement pnl for futures.
        type: number
      initial_margin:
        description: Initial margin requirement.
        type: number
      maintenance_margin:
        description: Maintenance margin requirement.
        type: number
      margin_call:
        description: Current margin call amount.
        type: number
      option_open_position:
        description: Number of open option contracts.
        type: number
      option_openbuy_market_value:
        description: Value of long option positions.
        type: number
      option_opensell_market_value:
        description: Value of short option positions.
        type: number
      option_settle_profitloss:
        description: Settlement pnl for options.
        type: number
      order_margin_premium:
        description: Margin premium for active orders.
        type: number
      plus_margin:
        description: Excess margin.
        type: number
      plus_margin_indicator:
        description: Excess margin indicator.
        type: number
      risk_indicator:
        description: Risk indicator percentage.
        type: number
      royalty_revenue_expenditure:
        description: Net royalty from options.
        type: number
      security_collateral_amount:
        description: Security collateral value.
        type: number
      status:
        allOf:
        - $ref: '#/definitions/pb.FetchStatus'
        description: Success or failure status.
      tax:
        description: Total taxes paid.
        type: number
      today_balance:
        description: Today's current balance.
        type: number
      today_future_open_position:
        description: Futures opened today.
        type: number
      yesterday_balance:
        description: Yesterday's ending balance.
        type: number
    type: object
  pb.Notice:
    properties:
      announced_date:
        description: Date of notice.
        items:
          type: string
        type: array
      close:
        description: Triggering close price.
        items:
          type: number
        type: array
      code:
        description: Stock codes.
        items:
          type: string
        type: array
      reason:
        description: Reason for notice.
        items:
          type: string
        type: array
      updated_at:
        description: Timestamp.
        items:
          type: string
        type: array
    type: object
...
  pb.Punish:
    properties:
      announced_date:
        description: Announcement date.
        items:
          type: string
        type: array
      code:
        description: Stock codes.
        items:
          type: string
        type: array
      description:
        description: Detail description.
        items:
          type: string
        type: array
      end_date:
        description: Punishment end.
        items:
          type: string
        type: array
      interval:
        description: Trade interval.
        items:
          type: string
        type: array
      start_date:
        description: Punishment start.
        items:
          type: string
        type: array
      total_limit:
        description: Total limit.
        items:
          type: number
        type: array
      unit_limit:
        description: Order unit limit.
        items:
          type: number
        type: array
      updated_at:
        description: Updated timestamp.
        items:
          type: string
        type: array
    type: object
...
  pb.ScannerType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    format: int32
    type: integer
    x-enum-comments:
      ScannerType_SCANNER_TYPE_AMOUNTRANK: Amount rank. 成交金額排行
      ScannerType_SCANNER_TYPE_CHANGEPERCENTRANK: Change percent rank. 漲跌幅排行
      ScannerType_SCANNER_TYPE_CHANGEPRICERANK: Change price rank. 漲跌價排行
      ScannerType_SCANNER_TYPE_DAYRANGERANK: Day range rank. 震幅排行
      ScannerType_SCANNER_TYPE_TICKCOUNTRANK: Tick count rank. 成交筆數排行
      ScannerType_SCANNER_TYPE_VOLUMERANK: Volume rank. 成交量排行
    x-enum-descriptions:
    - ""
    - Change percent rank. 漲跌幅排行
    - Change price rank. 漲跌價排行
    - Day range rank. 震幅排行
    - Volume rank. 成交量排行
    - Amount rank. 成交金額排行
    - Tick count rank. 成交筆數排行
    x-enum-varnames:
    - ScannerType_SCANNER_TYPE_UNSPECIFIED
    - ScannerType_SCANNER_TYPE_CHANGEPERCENTRANK
    - ScannerType_SCANNER_TYPE_CHANGEPRICERANK
    - ScannerType_SCANNER_TYPE_DAYRANGERANK
    - ScannerType_SCANNER_TYPE_VOLUMERANK
    - ScannerType_SCANNER_TYPE_AMOUNTRANK
    - ScannerType_SCANNER_TYPE_TICKCOUNTRANK
...
  pb.Settlement:
    properties:
      amount:
        description: Total amount.
        type: number
      date:
        description: Settlement date.
        type: string
      status:
        allOf:
        - $ref: '#/definitions/pb.FetchStatus'
        description: Fetch status.
      t_day:
        description: T date.
        type: string
      t_money:
        description: Amount for T day.
        type: number
      t1_day:
        description: T+1 date.
        type: string
      t1_money:
        description: Amount for T+1 day.
        type: number
      t2_day:
        description: T+2 date.
        type: string
      t2_money:
        description: Amount for T+2 day.
        type: number
    type: object
...
  pb.Ticks:
    properties:
      ask_price:
        description: Best ask at time of tick.
        items:
          type: number
        type: array
      ask_volume:
        description: Best ask volume.
        items:
          type: integer
        type: array
      bid_price:
        description: Best bid at time of tick.
        items:
          type: number
        type: array
      bid_volume:
        description: Best bid volume.
        items:
          type: integer
        type: array
      close:
        description: Execution prices.
        items:
          type: number
        type: array
      tick_type:
        description: Tick type indicators.
        items:
          type: integer
        type: array
      ts:
        description: Timestamps.
        items:
          type: integer
        type: array
      volume:
        description: Execution volumes.
        items:
          type: integer
        type: array
    type: object
...
  pb.TradingLimits:
    properties:
      margin_available:
        description: Remaining margin capacity.
        type: integer
      margin_limit:
        description: Total margin trading limit.
        type: integer
      margin_used:
        description: Utilized margin limit.
        type: integer
      short_available:
        description: Remaining short selling capacity.
        type: integer
      short_limit:
        description: Total short selling limit.
        type: integer
      short_used:
        description: Utilized short selling limit.
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/pb.FetchStatus'
        description: Fetch status.
      trading_available:
        description: Remaining trading capacity.
        type: integer
      trading_limit:
        description: Total allowed trading limit.
        type: integer
      trading_used:
        description: Currently utilized limit.
        type: integer
    type: object
...
  pb.UsageStatus:
    properties:
      bytes:
        description: Amount of data consumed in bytes.
        type: integer
      connections:
        description: Current number of active connections.
        type: integer
      limit_bytes:
        description: Maximum data consumption limit in bytes.
        type: integer
      remaining_bytes:
        description: Remaining data allowance in bytes.
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API Gateway for Shioaji Trading Provider.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Phoenix API
  version: "1.0"
paths:
  /api/v1/accounts:
    get:
      description: List all accounts available for the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.ListAccountsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: List Accounts
      tags:
      - Account
  /api/v1/accounts/balance:
    get:
      description: Get balance for the account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.AccountBalance'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Account Balance
      tags:
      - Account
  /api/v1/accounts/limits:
    get:
      consumes:
      - application/json
      description: Get trading limits information
      parameters:
      - description: Trading Limits Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.GetTradingLimitsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.TradingLimits'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Trading Limits
      tags:
      - Account
  /api/v1/accounts/margin:
    get:
      consumes:
      - application/json
      description: Get margin information
      parameters:
      - description: Margin Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.GetMarginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.Margin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Margin
      tags:
      - Account
  /api/v1/accounts/settlements:
    get:
      consumes:
      - application/json
      description: Get settlement information
      parameters:
      - description: Settlement Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.GetSettlementsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.GetSettlementsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Settlements
      tags:
      - Account
...
  /api/v1/market/contracts/fetch:
    post:
      consumes:
      - application/json
      description: Fetch contract definitions
      parameters:
      - description: Fetch Contracts Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.FetchContractsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.Empty'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Fetch Contracts
      tags:
      - Market
  /api/v1/market/daily-quotes:
    post:
      consumes:
      - application/json
      description: Get daily quotes
      parameters:
      - description: Daily Quotes Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.GetDailyQuotesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.DailyQuotes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Daily Quotes
      tags:
      - Market
  /api/v1/market/kbars:
    post:
      consumes:
      - application/json
      description: Get market kbars
      parameters:
      - description: Kbars Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.GetKbarsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.Kbars'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Kbars
      tags:
      - Market
  /api/v1/market/notice:
    get:
      description: Get market notices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.Notice'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Market Notices
      tags:
      - Market
  /api/v1/market/punish:
    get:
      description: Get stock punishment information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.Punish'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Punish Info
      tags:
      - Market
  /api/v1/market/scanners:
    post:
      consumes:
      - application/json
      description: Get market scanners
      parameters:
      - description: Scanners Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.GetScannersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.GetScannersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Scanners
      tags:
      - Market
  /api/v1/market/snapshots:
    post:
      consumes:
      - application/json
      description: Get market snapshots
      parameters:
      - description: Snapshots Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.GetSnapshotsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.GetSnapshotsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Snapshots
      tags:
      - Market
  /api/v1/market/ticks:
    post:
      consumes:
      - application/json
      description: Get market ticks
      parameters:
      - description: Ticks Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.GetTicksRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.Ticks'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Ticks
      tags:
      - Market
...
  /api/v1/usage:
    get:
      description: Get usage information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.UsageStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Usage
      tags:
      - Account
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"